+++ src/main/java/org/fundacionjala/pivotal/EndpointHelper.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.coding;
+package org.fundacionjala.pivotal;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
