--- src/test/java/org/fundacionjala/coding/tdd/POSTStoryTest.java
@@ -1,43 +1,0 @@
-package org.fundacionjala.coding.tdd;
-
-import io.restassured.response.Response;
-import org.fundacionjala.coding.RequestManager;
-import org.fundacionjala.coding.RequestSpecFactory;
-import org.testng.Assert;
-import org.testng.annotations.AfterTest;
-import org.testng.annotations.BeforeTest;
-import org.testng.annotations.Test;
-
-public class POSTStoryTest {
-
-    private String projectId;
-
-    @BeforeTest
-    public void setUp() {
-        //Given
-        String expectedProjectName = "Test Story project";
-        Response response = RequestManager.post(RequestSpecFactory.getRequestSpec("pivotal"),
-                "/projects",
-                "{\"name\":\"" + expectedProjectName + "\"}");
-        projectId = response.jsonPath().getString("id");
-    }
-
-    @Test
-    public void testPOSTStory() {
-        //When
-        String expectedStoryString = "story string";
-        Response response = RequestManager.post(RequestSpecFactory.getRequestSpec("pivotal"),
-                String.format("/projects/%s/stories", projectId),
-                "{\"name\":\"" + expectedStoryString + "\"}");
-
-        //Then
-        String actualStoryString = response.jsonPath().getString("name");
-        Assert.assertEquals(actualStoryString, expectedStoryString);
-    }
-
-    @AfterTest
-    public void cleanData() {
-        RequestManager.delete(RequestSpecFactory.getRequestSpec("pivotal"),
-                String.format("/projects/%s", projectId));
-    }
-}
