+++ src/test/java/org/fundacionjala/pivotal/runner/Runner.java
@@ -1,0 +1,11 @@
+package org.fundacionjala.pivotal.runner;
+
+import io.cucumber.testng.AbstractTestNGCucumberTests;
+import io.cucumber.testng.CucumberOptions;
+
+@CucumberOptions(
+        glue = {"org.fundacionjala.pivotal"},
+        features = "src/test/resources/features"
+)
+public class Runner extends AbstractTestNGCucumberTests {
+}
