+++ src/test/java/org/fundacionjala/pivotal/tdd/DELETEStoryCommentTest.java
@@ -1,16 +1,19 @@
-package org.fundacionjala.coding.tdd;
+package org.fundacionjala.pivotal.tdd;
 
 import io.restassured.response.Response;
-import org.fundacionjala.coding.RequestManager;
-import org.fundacionjala.coding.RequestSpecFactory;
+import org.fundacionjala.pivotal.RequestManager;
+import org.fundacionjala.pivotal.RequestSpecFactory;
 import org.testng.Assert;
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
-public class GETStoryTest {
+public class DELETEStoryCommentTest {
 
     private String projectId;
+    private String storyId;
+    private String storyCommentId;
+    private int statusCode;
 
     @BeforeTest
     public void setUp() {
@@ -26,18 +29,26 @@ public class GETStoryTest {
         response = RequestManager.post(RequestSpecFactory.getRequestSpec("pivotal"),
                 String.format("/projects/%s/stories", projectId),
                 "{\"name\":\"" + expectedStoryString + "\"}");
+        storyId = response.jsonPath().getString("id");
+
+        //And
+        String expectedStoryCommentString = "story comment";
+        response = RequestManager.post(RequestSpecFactory.getRequestSpec("pivotal"),
+                String.format("/projects/%s/stories/%s/comments", projectId, storyId),
+                "{\"text\":\"" + expectedStoryCommentString + "\"}");
+        storyCommentId = response.jsonPath().getString("id");
     }
 
     @Test
-    public void testGETStory() {
+    public void testDELETEStoryComment() {
         //When
-        String expectedStoryString = "[story string]";
-        Response response = RequestManager.get(RequestSpecFactory.getRequestSpec("pivotal"),
-                String.format("/projects/%s/stories/", projectId));
+        String expectedStoryCommentString = "story comment";
+        Response response = RequestManager.delete(RequestSpecFactory.getRequestSpec("pivotal"),
+                String.format("/projects/%s/stories/%s/comments/%s", projectId, storyId, storyCommentId));
 
         //Then
-        String actualStoryString = response.jsonPath().getString("name");
-        Assert.assertEquals(actualStoryString, expectedStoryString);
+        statusCode = response.getStatusCode();
+        Assert.assertEquals(statusCode, 204);
     }
 
     @AfterTest
