+++ src/test/java/org/fundacionjala/pivotal/tdd/DELETEStoryTest.java
@@ -1,18 +1,18 @@
-package org.fundacionjala.coding.tdd;
+package org.fundacionjala.pivotal.tdd;
 
 import io.restassured.response.Response;
-import org.fundacionjala.coding.RequestManager;
-import org.fundacionjala.coding.RequestSpecFactory;
+import org.fundacionjala.pivotal.RequestManager;
+import org.fundacionjala.pivotal.RequestSpecFactory;
 import org.testng.Assert;
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
-public class PUTStoryCommentTest {
+public class DELETEStoryTest {
 
     private String projectId;
     private String storyId;
-    private String storyCommentId;
+    private int statusCode;
 
     @BeforeTest
     public void setUp() {
@@ -29,26 +29,17 @@ public class PUTStoryCommentTest {
                 String.format("/projects/%s/stories", projectId),
                 "{\"name\":\"" + expectedStoryString + "\"}");
         storyId = response.jsonPath().getString("id");
-
-        //And
-        String expectedStoryCommentString = "story comment";
-        response = RequestManager.post(RequestSpecFactory.getRequestSpec("pivotal"),
-                String.format("/projects/%s/stories/%s/comments", projectId, storyId),
-                "{\"text\":\"" + expectedStoryCommentString + "\"}");
-        storyCommentId = response.jsonPath().getString("id");
     }
 
     @Test
-    public void testPUTStoryComment() {
+    public void testDELETEStory() {
         //When
-        String expectedStoryCommentString = "updated comment";
-        Response response = RequestManager.put(RequestSpecFactory.getRequestSpec("pivotal"),
-                String.format("/projects/%s/stories/%s/comments/%s", projectId, storyId, storyCommentId),
-                "{\"text\":\"" + expectedStoryCommentString + "\"}");
+        Response response = RequestManager.delete(RequestSpecFactory.getRequestSpec("pivotal"),
+                String.format("/projects/%s/stories/%s", projectId, storyId));
 
         //Then
-        String actualStoryCommentString = response.jsonPath().getString("text");
-        Assert.assertEquals(actualStoryCommentString, expectedStoryCommentString);
+        statusCode = response.getStatusCode();
+        Assert.assertEquals(statusCode, 204);
     }
 
     @AfterTest
