+++ src/test/java/org/fundacionjala/pivotal/tdd/GETStoryTest.java
@@ -1,18 +1,16 @@
-package org.fundacionjala.coding.tdd;
+package org.fundacionjala.pivotal.tdd;
 
 import io.restassured.response.Response;
-import org.fundacionjala.coding.RequestManager;
-import org.fundacionjala.coding.RequestSpecFactory;
+import org.fundacionjala.pivotal.RequestManager;
+import org.fundacionjala.pivotal.RequestSpecFactory;
 import org.testng.Assert;
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
-public class GETStoryCommentTest {
+public class GETStoryTest {
 
     private String projectId;
-    private String storyId;
-    private String storyCommentId;
 
     @BeforeTest
     public void setUp() {
@@ -28,26 +26,18 @@ public class GETStoryCommentTest {
         response = RequestManager.post(RequestSpecFactory.getRequestSpec("pivotal"),
                 String.format("/projects/%s/stories", projectId),
                 "{\"name\":\"" + expectedStoryString + "\"}");
-        storyId = response.jsonPath().getString("id");
-
-        //And
-        String expectedStoryCommentString = "story comment";
-        response = RequestManager.post(RequestSpecFactory.getRequestSpec("pivotal"),
-                String.format("/projects/%s/stories/%s/comments", projectId, storyId),
-                "{\"text\":\"" + expectedStoryCommentString + "\"}");
-        storyCommentId = response.jsonPath().getString("id");
     }
 
     @Test
-    public void testGETStoryComment() {
+    public void testGETStory() {
         //When
-        String expectedStoryCommentString = "story comment";
+        String expectedStoryString = "[story string]";
         Response response = RequestManager.get(RequestSpecFactory.getRequestSpec("pivotal"),
-                String.format("/projects/%s/stories/%s/comments/%s", projectId, storyId, storyCommentId));
+                String.format("/projects/%s/stories/", projectId));
 
         //Then
-        String actualStoryCommentString = response.jsonPath().getString("text");
-        Assert.assertEquals(actualStoryCommentString, expectedStoryCommentString);
+        String actualStoryString = response.jsonPath().getString("name");
+        Assert.assertEquals(actualStoryString, expectedStoryString);
     }
 
     @AfterTest
